[js/web] revise TensorView
